backend/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
segment-anything==1.0
torch>=1.7.0
torchvision>=0.8.0
aiohttp>=3.8.0
albumentations>=1.3.0
tensorboard>=2.12.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
psutil>=5.9.0
imageio>=2.31.1
pytest>=7.0.0
httpx>=0.25.2
